Move WHERE clause creation to query_constructors
