+++ tests/unit_tests/test_query_constructors.py
@@ -1,4 +1,4 @@
-from typing import Dict, Tuple
+from typing import Any, Dict, List, Tuple
 
 import pytest as pytest
 from langchain_core.structured_query import (
@@ -9,10 +9,23 @@ from langchain_core.structured_query import (
     StructuredQuery,
 )
 
-from langchain_hana.query_constructors import HanaTranslator
+from langchain_hana.query_constructors import CreateWhereClause, HanaTranslator
+from langchain_hana.vectorstores import default_metadata_column
+from tests.integration_tests.fixtures.filtering_test_cases import (
+    TYPE_1_FILTERING_TEST_CASES,
+    TYPE_2_FILTERING_TEST_CASES,
+    TYPE_3_FILTERING_TEST_CASES,
+    TYPE_4_FILTERING_TEST_CASES,
+    TYPE_4B_FILTERING_TEST_CASES,
+    TYPE_5_FILTERING_TEST_CASES,
+)
 
 DEFAULT_TRANSLATOR = HanaTranslator()
 
+class MockHanaDb:
+    def __init__(self):
+        self.metadata_column = default_metadata_column
+        self.specific_metadata_columns = []
 
 def test_visit_comparison() -> None:
     comp = Comparison(comparator=Comparator.LT, attribute="foo", value=1)
@@ -82,3 +95,86 @@ def test_visit_structured_query() -> None:
     )
     actual = DEFAULT_TRANSLATOR.visit_structured_query(structured_query)
     assert expected == actual
+
+def test_create_where_clause_empty_filter() -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())({})
+    assert where_clause == ""
+    assert parameters == []
+
+
+def test_create_where_clause_unexpected_operator() -> None:
+    invalid_filter = {"$eq": [{"key": "value"}]}
+    with pytest.raises(ValueError, match="Unexpected operator"):
+        CreateWhereClause(MockHanaDb())(invalid_filter)
+
+
+def test_create_where_clause_unsupported_filter_value_type() -> None:
+    unsupported_filter = {"key": [1, 2, 3]}
+    with pytest.raises(ValueError, match="Unsupported filter value type"):
+        CreateWhereClause(MockHanaDb())(unsupported_filter)
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_1_FILTERING_TEST_CASES)
+def test_create_where_clause_1(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_2_FILTERING_TEST_CASES)
+def test_create_where_clause_2(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_3_FILTERING_TEST_CASES)
+def test_create_where_clause_3(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_4_FILTERING_TEST_CASES)
+def test_create_where_clause_4(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_4B_FILTERING_TEST_CASES)
+def test_create_where_clause_4b(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
+
+@pytest.mark.parametrize("test_filter, expected_ids, expected_where_clause, expected_where_clause_parameters", TYPE_5_FILTERING_TEST_CASES)
+def test_create_where_clause_5(
+    test_filter: Dict[str, Any],
+    expected_ids: List[int],
+    expected_where_clause: str,
+    expected_where_clause_parameters: List[Any],
+) -> None:
+    where_clause, parameters = CreateWhereClause(MockHanaDb())(test_filter)
+    assert expected_where_clause == where_clause
+    assert expected_where_clause_parameters == parameters
